+++ src/TestSuite/Traits/TestCaseTrait.php
@@ -13,7 +13,8 @@
  */
 namespace MeTools\TestSuite\Traits;
 
-use Traversable;
+use Tools\ReflectionTrait;
+use Tools\TestSuite\TestCaseTrait as ToolsTestCaseTrait;
 
 /**
  * This trait provides some useful methods for `TestCase` and
@@ -21,155 +22,26 @@ use Traversable;
  */
 trait TestCaseTrait
 {
-    /**
-     * Asserts that the array keys are equal to `$expected`
-     * @param array $expected Expected keys
-     * @param array $array Array to check
-     * @param string $message The failure message that will be appended to the
-     *  generated message
-     * @return void
-     */
-    public function assertArrayKeysEqual($expected, $array, $message = '')
-    {
-        $this->assertIsArray($array);
-        $this->assertEquals($expected, array_keys($array), $message);
-    }
-
-    /**
-     * Asserts that a filename exists
-     * @param string|array|Traversable $filename Filename or array/Traversable
-     *  of filenames
-     * @param string $message The failure message that will be appended to the
-     *  generated message
-     * @return void
-     */
-    public static function assertFileExists($filename, $message = '')
-    {
-        if (is_array($filename) || $filename instanceof Traversable) {
-            foreach ($filename as $var) {
-                parent::assertFileExists($var, $message);
-            }
-
-            return;
-        }
-
-        parent::assertFileExists($filename, $message);
-    }
-
-    /**
-     * Asserts that a filename not exists
-     * @param string|array|Traversable $filename Filename or array/Traversable
-     *  of filenames
-     * @param string $message The failure message that will be appended to the
-     *  generated message
-     * @return void
-     */
-    public static function assertFileNotExists($filename, $message = '')
-    {
-        if (is_array($filename) || $filename instanceof Traversable) {
-            foreach ($filename as $var) {
-                parent::assertFileNotExists($var, $message);
-            }
-
-            return;
-        }
-
-        parent::assertFileNotExists($filename, $message);
-    }
-
-    /**
-     * Asserts that `$actual` is an instance of `$expected`
-     * @param string $expected Expected namespace
-     * @param array|Traversable|mixed $actual Instance or array/Traversable of
-     *  instances
-     * @param string $message The failure message that will be appended to the
-     *  generated message
-     * @return void
-     */
-    public static function assertInstanceOf($expected, $actual, $message = '')
-    {
-        if ((is_array($actual) || $actual instanceof Traversable) &&
-            !$actual instanceof \Cake\Validation\Validator
-        ) {
-            foreach ($actual as $var) {
-                parent::assertInstanceOf($expected, $var, $message);
-            }
-
-            return;
-        }
-
-        parent::assertInstanceOf($expected, $actual, $message);
-    }
-
-    /**
-     * Asserts that a variable is an array
-     * @param mixed $var Variable to check
-     * @param string $message The failure message that will be appended to the
-     *  generated message
-     * @return void
-     */
-    public function assertIsArray($var, $message = '')
-    {
-        $this->assertTrue(is_array($var), $message);
-    }
-
-    /**
-     * Asserts that a variable is an object
-     * @param mixed $var Variable to check
-     * @param string $message The failure message that will be appended to the
-     *  generated message
-     * @return void
-     */
-    public function assertIsObject($var, $message = '')
-    {
-        $this->assertTrue(is_object($var), $message);
-    }
-
-    /**
-     * Asserts that a variable is a string
-     * @param mixed $var Variable to check
-     * @param string $message The failure message that will be appended to the
-     *  generated message
-     * @return void
-     */
-    public function assertIsString($var, $message = '')
-    {
-        $this->assertTrue(is_string($var), $message);
-    }
+    use ReflectionTrait;
+    use ToolsTestCaseTrait;
 
     /**
      * Asserts log file contents
-     * @param string $expected The expected contents
-     * @param string $name Log name
+     * @param string $expectedContent The expected contents
+     * @param string $logName Log name
      * @param string $message The failure message that will be appended to the
      *  generated message
      * @return void
      */
-    public function assertLogContains($expected, $name, $message = '')
+    public function assertLogContains($expectedContent, $logName, $message = '')
     {
-        $file = LOGS . $name . '.log';
+        $file = LOGS . $logName . '.log';
 
         if (!is_readable($file)) {
-            $this->fail('Log file ' . $file . ' not readable');
+            $this->fail('Log file `' . $file . '` not readable');
         }
 
-        $content = trim(file_get_contents($file));
-
-        $this->assertContains($expected, $content, $message);
-    }
-
-    /**
-     * Asserts that the object properties are equal to `$expected`
-     * @param array $expected Expected keys
-     * @param array $object Ojbect to check
-     * @param string $message The failure message that will be appended to the
-     *  generated message
-     * @return void
-     */
-    public function assertObjectPropertiesEqual($expected, $object, $message = '')
-    {
-        $this->assertIsObject($object);
-        $this->assertEquals($expected, array_keys((array)$object), $message);
+        $this->assertContains($expectedContent, file_get_contents($file), $message);
     }
 
     /**
@@ -186,12 +58,12 @@ trait TestCaseTrait
 
     /**
      * Deletes a log file
-     * @param string $name Log name
+     * @param string $logName Log name
      * @return void
      */
-    public function deleteLog($name)
+    public function deleteLog($logName)
     {
         //@codingStandardsIgnoreLine
-        @unlink(LOGS . $name . '.log');
+        @unlink(LOGS . $logName . '.log');
     }
 }